Move priority_search under opto/features
