+++ examples/priority_search_example.py
@@ -3,7 +3,7 @@ import numpy as np
 from opto import trace
 from opto.utils.llm import LLM, LiteLLM
 from opto.optimizers import OptoPrimeV2 as OptoPrime
-from opto.trainer.algorithms.priority_search import PrioritySearch as SearchAlgorithm
+from opto.features.priority_search import PrioritySearch as SearchAlgorithm
 from opto.trainer.loggers import TensorboardLogger
 from opto.trainer.guide import VerbalJudgeGuide
 from typing import Any
