+++ opto/features/priority_search/__init__.py
@@ -1,0 +1,2 @@
+from opto.features.priority_search.priority_search import PrioritySearch
+from opto.features.priority_search.examples import SequentialUpdate, SequentialSearch, BeamSearch
